+++ contracts/stake/src/contract.rs
@@ -181,6 +181,14 @@ impl StakingTrait for Staking {
 
         let config = get_config(&env);
 
+        // check for rewards and withdraw them
+        let found_rewards: WithdrawableRewardsResponse =
+            Self::query_withdrawable_rewards(env.clone(), sender.clone());
+
+        if !found_rewards.rewards.is_empty() {
+            Self::withdraw_rewards(env.clone(), sender.clone());
+        }
+
         let mut stakes = get_stakes(&env, &sender);
         remove_stake(&mut stakes.stakes, stake_amount, stake_timestamp);
         stakes.total_stake -= stake_amount as u128;