+++ contracts/stake/src/tests/bond.rs
@@ -226,3 +226,53 @@ fn unbond_wrong_user_stake_not_found() {
 
     staking.unbond(&user2, &10_000, &2_000);
 }
+
+#[test]
+fn pay_rewards_during_unbond() {
+    const STAKED_AMOUNT: i128 = 1_000;
+    let env = Env::default();
+    env.mock_all_auths();
+
+    let admin = Address::generate(&env);
+    let user = Address::generate(&env);
+    let manager = Address::generate(&env);
+    let owner = Address::generate(&env);
+
+    let lp_token = deploy_token_contract(&env, &admin);
+    let reward_token = deploy_token_contract(&env, &admin);
+    let staking = deploy_staking_contract(&env, admin.clone(), &lp_token.address, &manager, &owner);
+
+    lp_token.mint(&user, &10_000);
+    reward_token.mint(&admin, &10_000);
+
+    staking.create_distribution_flow(&manager, &reward_token.address);
+    staking.fund_distribution(&admin, &0u64, &10_000u64, &reward_token.address, &10_000);
+
+    staking.bond(&user, &STAKED_AMOUNT);
+
+    env.ledger().with_mut(|li| {
+        li.timestamp = 5_000;
+    });
+    staking.distribute_rewards();
+
+    // user has bonded for 5_000 time, initial rewards are 10_000
+    // so user should have 5_000 rewards
+    // 5_000 rewards are still undistributed
+    assert_eq!(
+        staking.query_undistributed_rewards(&reward_token.address),
+        5_000
+    );
+    assert_eq!(
+        staking
+            .query_withdrawable_rewards(&user)
+            .rewards
+            .iter()
+            .map(|reward| reward.reward_amount)
+            .sum::<u128>(),
+        5_000
+    );
+
+    assert_eq!(reward_token.balance(&user), 0);
+    staking.unbond(&user, &STAKED_AMOUNT, &0);
+    assert_eq!(reward_token.balance(&user), 5_000);
+}