+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to
 
 ## Changed
 
+- Stake: unbonding now automatically pays the rewards ([#170])
 - Update soroban-sdk version from v20.0.3 to v20.1.0 ([#193])
 - Fixes documentation and naming ([#200])
 - Multihop: adds a new field in the Swap struct, that hold max_belief_price ([234])
@@ -20,6 +21,7 @@ and this project adheres to
 - Factory, Multihop, Pool, Pool_stable: adds a new parameter for creating liquidity pool ([#243])
 - Update soroban-sdk version from v20.1.0 to v20.4.0 ([#262])
 
+[#170]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/170
 [#200]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/200
 [#234]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/234
 [#235]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/235